refactor(credentials): centralize readiness input contract
