+++ packages/interfaces/src/publisher/index.ts
@@ -1,3 +1,4 @@
+export * from './credential-token-publishing-readiness-input.interface';
 export * from './publish-approval.interface';
 export * from './publisher.interface';
 export * from './publishing-readiness.interface';
